+++ backend/src/main/java/com/a804/tictactoc/ttt/response/HabitRes.java
@@ -24,7 +24,7 @@ public class HabitRes {
 //	long userId;
 
 	@Schema(name = "어떤 카테고리의 습관인지", example = "1", defaultValue = "1")
-	long categoryId;
+	int categoryId;
 
 	@Schema(name = "습관의 이름", example = "물마시기", defaultValue = "물마시기")
 	String name;
